Split final qualification policy (#451)
